+++ Modules/Intent.Modules.ApplicationTemplate.Builder/Model/ApplicationTemplateComponentDependency.cs
@@ -1,0 +1,13 @@
+﻿#nullable disable
+using System.Xml.Serialization;
+
+namespace Intent.Modules.ApplicationTemplate.Builder.Model;
+
+public class ApplicationTemplateComponentDependency
+{
+    [XmlAttribute("id")]
+    public string Id { get; set; }
+
+    [XmlAttribute("name")]
+    public string Name { get; set; }
+}