+++ Modules/Intent.Modules.ApplicationTemplate.Builder/Model/IconModelPersistable.cs
@@ -1,0 +1,15 @@
+#nullable disable
+using System.Xml.Serialization;
+using Intent.Metadata.Models;
+
+namespace Intent.Modules.ApplicationTemplate.Builder.Model
+{
+    public class IconModelPersistable
+    {
+        [XmlAttribute("type")]
+        public IconType Type { get; set; }
+
+        [XmlAttribute("source")]
+        public string Source { get; set; }
+    }
+}