+++ Modules/Intent.Modules.ApplicationTemplate.Builder/Model/MinimumDependencyVersion.cs
@@ -1,0 +1,13 @@
+﻿#nullable disable
+using System.Xml.Serialization;
+
+namespace Intent.Modules.ApplicationTemplate.Builder.Model;
+
+public class MinimumDependencyVersion
+{
+    [XmlAttribute("id")]
+    public string Id { get; set; }
+
+    [XmlAttribute("version")]
+    public string Version { get; set; }
+}