+++ Modules/Intent.Modules.ApplicationTemplate.Builder/Model/SettingFieldOptionPersistable.cs
@@ -1,0 +1,13 @@
+﻿#nullable disable
+using System.Xml.Serialization;
+
+namespace Intent.Modules.ApplicationTemplate.Builder.Model;
+
+public class SettingFieldOptionPersistable
+{
+    [XmlAttribute("value")]
+    public string Value { get; set; }
+
+    [XmlAttribute("description")]
+    public string Description { get; set; }
+}