+++ Modules/Intent.Modules.ApplicationTemplate.Builder/NuGet/Versioning/Resources.cs
@@ -1,0 +1,11 @@
+﻿namespace Intent.IArchitect.Common.NuGet.Versioning
+{
+    internal class Resources
+    {
+        public const string Argument_Cannot_Be_Null_Or_Empty = "Value cannot be null or an empty string.";
+        public const string Invalidvalue = "'{0}' is not a valid version string.";
+        public const string TypeNotSupported = "Type '{0}' is not supported.";
+        public const string CannotBeNullWhenParameterIsNotNull = "Parameter '{0}' cannot be null when parameter '{1}' is not null.";
+        public const string InvalidFloatRangeValue = "'{0}' is not a valid float range string.";
+    }
+}